// src/lib/types.ts
export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>;
// ...
export interface TblProjecten {
  ProjectAanbrengerID: string | null;
  ProjectCode: Generated<string | null>;
  ProjectContactID: string | null;
  ProjectDatumBevestiging: Date | null;
  ProjectDatumEind: Date | null;
  ProjectDatumOfferte: Date | null;
  ProjectDatumStart: Date | null;
  ProjectDeadline: Date | null;
  ProjectFacturatiePrijs: number | null;
  ProjectFolder: string | null;
  ProjectGewijzigdDoor: string | null;
  ProjectGewijzigdOp: Date | null;
  ProjectGUID: Generated<string>;
  ProjectLeiderID: string | null;
  ProjectMedewerker: string | null;
  ProjectMemoVerval: string | null;
  ProjectOmschrijving: string | null;
  ProjectOpmerking: string | null;
  ProjectReferteKlant: string | null;
  ProjectRegie: boolean | null;
  ProjectRelatieID: string | null;
  ProjectSoortID: string | null;
  ProjectStatusID: string | null;
  ProjectVervalID: string | null;
  ProjectVestigingID: string | null;
}